requirements.txt
Flask==1.1.1
pymongo==3.0.3
beautifulsoup4==4.4.1
requests==2.8.0
flask-socketio==4.2.1
eventlet==0.25.1
gunicorn==20.0.4
passlib==1.7.2
uwsgi
flask-login==0.5.0
flask-WTF==0.14.3
flask-Bootstrap==3.3.7.1
wtforms-validators==1.0.0
email-validator==1.0.5